Fix for issue #782 (https://github.com/django-tastypie/django-tastypie/issues/782)
Schema information only gives information about whether a related field is
'to_one' or 'to_many', but not about the actual resource type of the field. This
fix adds a new field, called 'related_schema' to the fields schema entry, which
defines the exact type of the related field:

    'user': {
        [..]
        'type': 'related'
        'related_type': 'to_one'
        'related_schema': '/api/v1/user/schema/'
        }

Documentation (schema example on the 'Interacting With The API' page) is updated accordingly.